polling-app/python/requirements.txt
PyQt6==6.4.2
python_bcrypt==0.3.2